+++ internal/core/port/glue.go
@@ -1,9 +1,0 @@
-package port
-
-import (
-	"context"
-)
-
-type GlueCodec interface {
-	Encode(ctx context.Context, schemaName string, payload []byte) ([]byte, error)
-}